+++ Mods ChoGGi/Spice Harvester/Code/SpiceHarvester.lua
@@ -9,7 +9,7 @@ function OnMsg.ModsReloaded()
 	local idx = table.find(ModsLoaded,"id","ChoGGi_Library")
 
 	if idx then
-		if library_version < ModsLoaded[idx].version then
+		if library_version > ModsLoaded[idx].version then
 			not_found_or_wrong_version = true
 		end
 	else
@@ -22,7 +22,7 @@ function OnMsg.ModsReloaded()
 			while not UICity do
 				Sleep(1000)
 			end
-			if WaitMarsQuestion(nil,nil,string.format([[Error: This mod requires ChoGGi's Library v%s.
+			if WaitMarsQuestion(nil,nil,string.format([[Error: This mod requires ChoGGi's Library (at least v%s).
 Press Ok to download it or check Mod Manager to make sure it's enabled.]],library_version)) == "ok" then
 				OpenUrl("https://steamcommunity.com/sharedfiles/filedetails/?id=1504386374")
 			end