+++ Mods ChoGGi/View Colony Map/metadata.lua
@@ -1,14 +1,16 @@
 return PlaceObj("ModDef", {
-	"title", "View Colony Map v0.4",
-	"version", 4,
-	"saved", 1536840000,
+	"title", "View Colony Map v0.5",
+	"version", 5,
+	"saved", 1538049600,
 	"image", "Preview.png",
 	"id", "ChoGGi_ViewColonyMap",
 	"steam_id", "1491973763",
 	"author", "ChoGGi",
-	"code", {"Code/Script.lua"},
+	"code", {
+		"Code/Script.lua"
+	},
 	"lua_revision", LuaRevision,
-	"description", [[Shows the map before you load it in the "Colony Spot" screen.
+	"description", [[Shows the map before you load it in the select colony screen.
 
 For those wondering: Purple = mountainous, other colours = buildable.